% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_eff.crd.R
\name{diff_eff.crd}
\alias{diff_eff.crd}
\title{Completely Randomized Design}
\usage{
diff_eff.crd(resp, trt, los = 0.05)
}
\arguments{
\item{resp}{The response variable vector}

\item{trt}{The treatment vector}

\item{los}{Level of significance (Default is 0.05)}
}
\value{
A list containing - ANOVA Table, Decision Table, Rejected Pairs, Mean Square Error, Critical Point
}
\description{
Test for Significant Treatment effects and Differential Effects of Treatment Contrasts for a CRD
}
\details{
In experimental designs if test of differential effect gets rejected we might be interested in analyzing which pair of treatments is behind the rejection.
For CRD we need to calculate the estimate of the treatment contrast (tau_i - tau_i') and test H0: tau_{i} = tau_{i'} for plausible rejection.
The estimate of (tau_i - tau_i') is (y_{i0} - y_{i'0}) which follows N(tau_i - tau_i', sigma^2 * (1/n_i + 1/n_i')).
Under H_0, the test statistic (y_{i0} - y_{i'0})/sqrt(MSE*(1/n_i + 1/n_i')) follows t_{n-v}.
}
\section{Special Thanks}{
 Professor Surupa Chakraborty and Professor Debjit Sengupta for helping me in building the concepts of Design of Experiments.
Professor Madhura Dasgupta for guiding me in R programming.
}

\seealso{
For RBD \code{\link{diff_eff.rbd}}, for LSD \code{\link{diff_eff.lsd}}
}
\author{
Anik Chakraborty
}
