% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_eff.lsd.R
\name{diff_eff.lsd}
\alias{diff_eff.lsd}
\title{Latin Square Design}
\usage{
diff_eff.lsd(resp, trt, rows, cols, los = 0.05)
}
\arguments{
\item{resp}{The response variable vector}

\item{trt}{The treatment vector}

\item{rows}{A vector containing the row indices of a LSD}

\item{cols}{A vector containing the column indices of a LSD}

\item{los}{Level of significance (Default is 0.05)}
}
\value{
A list containing - ANOVA Table, Decision Table, Rejected Pairs, Mean Square Error, Critical Point
}
\description{
Test for Significant Treatment effects and Differential Effects of Treatment Contrasts for a LSD
}
\details{
In experimental designs if test of differential effect gets rejected we might be interested in analyzing which pair of treatments is behind the rejection.
For LSD we need to calculate the estimate of the treatment contrast (tau_k - tau_k') and test H0: tau_{k} = tau_{k'} for plausible rejection.
The estimate of (tau_k - tau_k') is (y_{00k} - y_{00k'}) which follows N(tau_j - tau_j', 2*sigma^2/v).
Under H_0, the test statistic (y_{00k} - y_{00k'})/sqrt(2*MSE/v) follows t_{(v-1)(v-2)}.
}
\section{Special Thanks}{
 Professor Surupa Chakraborty and Professor Debjit Sengupta for helping me in building the concepts of Design of Experiments.
Professor Madhura Dasgupta for guiding me in R programming.
}

\seealso{
For CRD \code{\link{diff_eff.crd}}, for RBD \code{\link{diff_eff.rbd}}
}
\author{
Anik Chakraborty
}
