% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_eff.rbd.R
\name{diff_eff.rbd}
\alias{diff_eff.rbd}
\title{Randomized Block Design}
\usage{
diff_eff.rbd(resp, trt, blocks, los = 0.05)
}
\arguments{
\item{resp}{The response variable vector}

\item{trt}{The treatment vector}

\item{blocks}{The vector containing Block/Replicate indices}

\item{los}{Level of significance (Default is 0.05)}
}
\value{
A list containing - ANOVA Table, Decision Table, Rejected Pairs, Mean Square Error, Critical Point
}
\description{
Test for Significant Treatment effects and Differential Effects of Treatment Contrasts for an RBD
}
\details{
In experimental designs if test of differential effect gets rejected we might be interested in analyzing which pair of treatments is behind the rejection.
For RBD we need to calculate the estimate of the treatment contrast (tau_j - tau_j') and test H0: tau_{j} = tau_{j'} for plausible rejection.
The estimate of (tau_j - tau_j') is (y_{0j} - y_{0j'}) which follows N(tau_j - tau_j', 2*sigma^2/b).
Under H_0, the test statistic (y_{0j} - y_{0j'})/sqrt(2*MSE/b) follows t_{(b-1)(v-1)}.
}
\section{Special Thanks}{
 Professor Surupa Chakraborty and Professor Debjit Sengupta for helping me in building the concepts of Design of Experiments.
Professor Madhura Dasgupta for guiding me in R programming.
}

\seealso{
For CRD \code{\link{diff_eff.crd}}, for LSD \code{\link{diff_eff.lsd}}
}
\author{
Anik Chakraborty
}
